Add gtk-dark.css to Dark and Darker themes.
This fixes libadwaita/libhandy apps which support dark mode.

Unfortunately this fix can't be applied to the light theme. In
MATE we're still using metacity for the titlebar. This theme
is legacy and doesn't have support for dynamic toolbar color.

In other words the light theme can't yet support dark apps.